+++ packages/com.eframework.debug-console/Plugins/IngameDebugConsole/Scripts/DebugLogConsole.cs
@@ -146,7 +146,6 @@ namespace IngameDebugConsole
 				"Mono.",
 				"mscorlib",
 				"netstandard",
-				"TextMeshPro",
 				"Microsoft.GeneratedCode",
 				"I18N",
 				"Boo.",