+++ packages/com.eframework.debug-console/Plugins/IngameDebugConsole/Scripts/DebugLogPopup.cs
@@ -2,7 +2,6 @@
 using UnityEngine.UI;
 using UnityEngine.EventSystems;
 using System.Collections;
-using TMPro;
 #if UNITY_EDITOR && UNITY_2021_1_OR_NEWER
 using Screen = UnityEngine.Device.Screen; // To support Device Simulator on Unity 2021.1+
 #endif
@@ -28,11 +27,11 @@ namespace IngameDebugConsole
 		private DebugLogManager debugManager;
 
 		[SerializeField]
-		private TextMeshProUGUI newInfoCountText;
+		private Text newInfoCountText;
 		[SerializeField]
-		private TextMeshProUGUI newWarningCountText;
+		private Text newWarningCountText;
 		[SerializeField]
-		private TextMeshProUGUI newErrorCountText;
+		private Text newErrorCountText;
 
 		[SerializeField]
 		private Color alertColorInfo;